Update GoogleMobileAds Framework
An error happened regarding the GoogleMobileAds framework and, after reinstalling it, it fixed itself.

Removed multiple 'non-existant' views from the storyboard. When creating a view on any view controller, upon deletion, ensure all the links to it are also removed (from the view controller actions, for example).

Updated some code in order to comply with Swift 2.